backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
PyMySQL>=1.0
mysqlclient>=2.1
python-dotenv>=1.0
Flask-Limiter>=2.0
gunicorn>=20.0
Flask-JWT-Extended>=4.4
Flask-Cors>=3.0
marshmallow>=3.0
bcrypt>=4.0
Flask-Migrate>=4.0
face_recognition>=1.3.0
numpy>=1.23
